process-images-from-cli/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
toml==0.10.2
